feat(settings): configure template directory path to include 'templates' folder
This commit updates the Django settings for the application to specify the location of templates. By including the BASE_DIR in combination with a subdirectory named 'templates', our Django project now knows where to look for template files. This is a conventional setup which facilitates the organization of template files within a dedicated directory, making it clearer and more maintainable.